(fix) Move HtmlFormEntryForm type to patient-common-lib
